Add files via upload
A base model for the 'Clustering BioTrove' machine learning challenge from the Biotrove_1 team. This is the second base model from the team. Whereas the original model extracted features from a single layer of the ResNet 50 convolutional neural network (CNN), this model extracts features from two separate levels of the ResNet 50 CNN, which allows for more general features to be extracted for genus-level clusters, and more specific features to be extracted for species-level clusters. Then, each set of features are input into a K-Means MiniBatch clustering algorithm.